Clean up the moxfield account linking in settings modal. refactor settings modal
